+++ SportAPI/Sport/Services/Interfaces/ITrainingService.cs
@@ -8,10 +8,10 @@ namespace SportAPI.Sport.Services.Interfaces
 {
   public interface ITrainingService
   {
-    Task<TrainingDto> GetById(int id);
+    Task<TrainingDto> GetById(long id);
     Task<IEnumerable<TrainingDto>> GetAll();
-    Task<int> Create(TrainingDto dto);
-    Task Delete(int id);
-    Task Update(int id, TrainingDto dto);
+    Task<long> Create(TrainingDto dto);
+    Task Delete(long id);
+    Task Update(long id, TrainingDto dto);
   }
 }
